requirements.txt
numpy>=1.21.0
tensorflow>=2.9.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
Flask>=2.0.0
Pillow>=8.0.0
h5py>=3.1.0
hbdscan>=0.8.27